Updated shop entities

+++ src/main/java/ru/ithub/practise/shop/entity/Milk.java
@@ -1,7 +1,7 @@
 package ru.ithub.practise.shop.entity;
 
 public class Milk extends Product {
-    public Milk(Integer amount) {
+    public Milk(Double amount) {
         super("Молоко", ProductCostType.LITER, 80.0, amount);
     }
 }
